requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosmtplib==1.1.6
anyio==3.6.1
Arpeggio==1.10.2
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
boto3==1.24.42
botocore==1.27.42
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
fastapi-mail==1.1.4
faunadb==4.2.0
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
grpcio==1.47.0
h11==0.13.0
h2==2.6.2
hpack==3.0.0
httptools==0.4.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
names==0.3.0
orjson==3.7.11
parver==0.3.1
protobuf==4.21.4
psycopg2-binary==2.9.3
pulumi==3.37.2
pulumi-aws==5.10.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytube==12.1.0
PyYAML==5.4.1
requests==2.28.1
rich==12.5.1
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
yarl==1.7.2
